Update socket service and add socket disconnect on sign out
