+++ src/client/services/signalR.service.js
@@ -1,45 +1,38 @@
 // If you use this starter with .NET back-end then
 // follow the instructions below, otherwise delete this file.
 
-// 1. Uninstall `socket.io-client` and delete `socketIo.service.js`
-// 2. Install `@microsoft/signalr`
-// 3. Change imports from `socketIo.service` to `signalR.service`
+// 1. Uninstall `socket.io-client`
+// 2. Delete `socket.service.js`
+// 3. Install `@microsoft/signalr`
+// 4. Rename this file to `socket.service.js`
 
-import config from 'config';
-import * as userHandlers from 'resources/user/user.handlers';
 import * as signalR from '@microsoft/signalr'; // eslint-disable-line import/no-unresolved
 
-let connection;
+import config from 'config';
 
-export const emit = (event, ...args) => {
-  if (!connection) return;
-  connection.invoke(event, ...args);
-};
+const connection = new signalR.HubConnectionBuilder()
+  .withUrl(config.webSocketUrl)
+  .withAutomaticReconnect()
+  .build();
 
-export const on = (event, callback) => {
-  if (!connection) return;
-  connection.on(event, callback);
-};
-
-export const connect = async () => {
-  connection = new signalR.HubConnectionBuilder()
-    .withUrl(config.webSocketUrl)
-    .withAutomaticReconnect()
-    .build();
-
-  await connection.start();
-
-  console.log('WS connected!'); // eslint-disable-line no-console
-  userHandlers.attachSocketEvents({ on, emit });
+export const connect = () => {
+  connection.start();
 };
 
 export const disconnect = () => {
-  if (!connection) return;
   connection.stop();
 };
 
+export const send = (event, ...args) => {
+  connection.send(event, ...args);
+};
+
+export const on = (event, callback) => {
+  connection.on(event, callback);
+};
+
 export const connected = () => {
-  return connection && connection.state === signalR.HubConnectionState.Connected;
+  return connection.state === signalR.HubConnectionState.Connected;
 };
 
 export const disconnected = () => {
